+++ src/components/AsyncProjects.js
@@ -1,5 +1,5 @@
 import { asyncComponent } from 'react-async-component';
 
 export default asyncComponent({
-  resolve: () => import('./Projects')
+  resolve: () => import('./Projects'),
 });