+++ src/components/footer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const Footer = () => (<div className="row">
   <div className="col-lg-12">
-    <br/>
+    <br />
       ©2009~2017 by <a href="http://www.gasolin.idv.tw">gasolin</a>, hosted on <a href="https://github.com/gasolin/gasolin.github.io">github</a>
   </div>
 </div>);