+++ src/index.js
@@ -5,6 +5,8 @@ import { AppContainer } from 'react-hot-loader';
 import { BrowserRouter } from 'react-router-dom';
 import 'material-components-web/dist/material-components-web.min.css';
 import App from './App';
+import './static/js/redirect';
+import './static/js/webfont';
 
 const renderApp = () => {
   render(
@@ -16,50 +18,6 @@ const renderApp = () => {
     document.getElementById('root'),
   );
 };
-
-// Single Page Apps for GitHub Pages
-// https://github.com/rafrex/spa-github-pages
-// Copyright (c) 2016 Rafael Pedicini, licensed under the MIT License
-// ----------------------------------------------------------------------
-// This script checks to see if a redirect is present in the query string
-// and converts it back into the correct url and adds it to the
-// browser's history using window.history.replaceState(...),
-// which won't cause the browser to attempt to load the new url.
-// When the single page app is loaded further down in this file,
-// the correct url will be waiting in the browser's history for
-// the single page app to route accordingly.
-(function(l) { // eslint-ignore func-names
-  if (l.search) {
-    var q = {};
-    l.search.slice(1).split('&').forEach(function(v) {
-      var a = v.split('=');
-      q[a[0]] = a.slice(1).join('=').replace(/~and~/g, '&');
-    });
-    if (q.p !== undefined) {
-      window.history.replaceState(null, null,
-        l.pathname.slice(0, -1) + (q.p || '') +
-        (q.q ? ('?' + q.q) : '') +
-        l.hash
-      );
-    }
-  }
-}(window.location))
-
-// This script adds the Roboto font to our project. For more detail go to this site:  http://www.google.com/fonts#UsePlace:use/Collection:Roboto:400,300,500 -->
-const WebFontConfig = {  // eslint-ignore no-unused-vars
-  google: {
-    families: ['Roboto:400,300,500:latin'],
-  },
-};
-
-(function () { // eslint-ignore func-names
-  const wf = document.createElement('script');
-  wf.src = 'https://ajax.googleapis.com/ajax/libs/webfont/1/webfont.js';
-  wf.type = 'text/javascript';
-  wf.async = 'true';
-  const s = document.getElementsByTagName('script')[0];
-  s.parentNode.insertBefore(wf, s);
-}());
 
 // This is needed for Hot Module Replacement
 if (module.hot) {